Improved Raspberry bash support

+++ scripts/install.sh
@@ -6,12 +6,12 @@
 # There's no need to handle version upgrade as any plugin update will trigger a full reinstallation
 #
 
-if [[ $# -ge 1 ]] && [[ -n $1 ]]; then
+if [ $# -ge 1 ] && [ -n $1 ]; then
 	virtualEnvFolder="$1"
 	/bin/echo "Installing into destination $virtualEnvFolder"
 else
 	isDarwinOS=$(/usr/bin/uname -a | /usr/bin/grep -c "Darwin")
-	if [[ $isDarwinOS -eq 1 ]]; then
+	if [ $isDarwinOS -eq 1 ]; then
 		virtualEnvFolder="/usr/local/lib/node_modules/homebridge-pichler-pkom4/scripts"
 		/bin/echo "Installing into default macOS destination"
 	else
@@ -26,13 +26,13 @@ pipPath="$virtualEnvFolder/bin/pip"
 modbusInstalled=0
 venvInstalled=$(/usr/bin/command -v $pipPath 2>&1 | /usr/bin/grep -c "/bin/pip")
 
-if [[ $venvInstalled -eq 1 ]]; then
+if [ $venvInstalled -eq 1 ]; then
 	/bin/echo "Checking available python virtual environment"	
 	modbusInstalled=$($pipPath show minimalmodbus 2>&1 | /usr/bin/grep -c "Version:")
 fi
 
 # Deal with possibly corrupted environment
-if [[ $venvInstalled -eq 1 ]] && [[ $modbusInstalled -eq 0 ]]; then
+if [ $venvInstalled -eq 1 ] && [ $modbusInstalled -eq 0 ]; then
 	/bin/echo "Removing previous python virtual environment"
 	venvInstalled=0
 	
@@ -42,21 +42,21 @@ if [[ $venvInstalled -eq 1 ]] && [[ $modbusInstalled -eq 0 ]]; then
 fi
 
 # Create virtual python env.
-if [[ $venvInstalled -eq 0 ]]; then
+if [ $venvInstalled -eq 0 ]; then
 	/bin/echo "Creating private python virtual environment"
 	/usr/bin/python3 -m venv "$virtualEnvFolder"
 fi
 
 # Install dependencies
-if [[ -f $pipPath ]] && [[ $modbusInstalled -eq 0 ]]; then
+if [ -f $pipPath ] && [ $modbusInstalled -eq 0 ]; then
 	/bin/echo "Installing minimal modbus"
 	$pipPath install minimalmodbus
 	
 	modbusInstalled=$($pipPath show minimalmodbus | /usr/bin/grep -c "Version:")
-	if [[ $modbusInstalled -eq 0 ]]; then
+	if [ $modbusInstalled -eq 0 ]; then
 		/bin/echo "Error: Failed to install minimal modbus"
 	fi
-elif [[ -f $pipPath ]]; then
+elif [ -f $pipPath ]; then
 	/bin/echo "Completed installation"
 else
 	/bin/echo "Error: Failed to install python virtual environment"
